ref(account): store simplify and remove account chains and balance assets
