--- app/stores/account/types.ts
@@ -1,26 +1,0 @@
-import type { Chain } from '@/types/chain'
-
-export type SubTokenKey = 'DOT' | 'KSM'
-export type EvmTokenKey = 'ETH' | 'WND'
-export type TokenKey = SubTokenKey | EvmTokenKey
-
-export interface TokenDetail {
-  balance: string
-  nativeBalance: string
-}
-
-export interface ChainData {
-  address: string
-  assets: Partial<Record<TokenKey, TokenDetail>>
-}
-
-export interface AccountVm {
-  address: string
-  chains: Partial<Record<Chain, ChainData>>
-  authSignature?: string
-}
-
-export interface SupportedAsset {
-  chain: Chain
-  token: TokenKey
-}
